Version 0.5 release

+++ Scancess/Scancess/Form1.cs
@@ -6,17 +6,24 @@ using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
+using System.Timers;
 using System.Windows.Forms;
 
 namespace Scancess
 {
     public partial class Form1 : Form
     {
+        private bool securemode = true;
+        private int devcount = 0;
+
         private DateTime _dt = DateTime.Now;  //定义一个成员函数用于保存每次的时间点
+        private string currentFileName = "fzsdev-1.csv";
         public Form1()
         {
             InitializeComponent();
+            textBox10.Text = getCSVcontent(currentFileName);
         }
 
         private void Form1_Load(object sender, EventArgs e)
@@ -65,28 +72,70 @@ namespace Scancess
                 //清空textbox
                 textBox1.Text = "";
 
-                //CSVFileHelper.SaveCSV(dt,"e:\\cs.csv");
-                string newFileName = "E:\\1.csv";
+                string newFileName = currentFileName;
                 string clientDetails = mac + "," + ssid + "," + aa + "," + bb + "," + cc + "," + dd + "," + apname + "," + appw;
+
+                if (getCSVcontent(newFileName).Contains(clientDetails))
+                {
+                    MessageBox.Show("设备信息已存在，无法保存！");
+                    return;
+                }
+
                 if (!File.Exists(newFileName))
                 {
-                    string clientHeader = "mac,ssid,aa,bb,cc,dd,apname,appw"+"\r\n";
+                    string clientHeader = "MAC,SSID,IMEI0,IMSI,ICCID,devSn,ApName,ApPwd"+"\r\n";
                     File.WriteAllText(newFileName, clientHeader);
                 }
                 try
                 {
+                    devcount++;
+                    if (securemode == true && devcount > 5)
+                        return;
+                    
                     File.AppendAllText(newFileName, clientDetails);
                 }
                 catch (Exception ex)
                 {
                     MessageBox.Show("CSV文件写入失败！error code="+ex);
                 }
+
+                textBox10.Text = getCSVcontent(currentFileName);
             }
         }
 
         private void textBox2_TextChanged(object sender, EventArgs e)
         {
 
+        }
+
+        private void 关于ToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            MessageBox.Show("Scancess|版本v0.5\n2020-11-10");
+        }
+
+        private void 保存路径ToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void textBox10_TextChanged(object sender, EventArgs e)
+        {
+            textBox10.Text = getCSVcontent(currentFileName);
+        }
+
+        private string getCSVcontent(string filePath) {
+            if (File.Exists(filePath))
+            {
+                try
+                {
+                    return File.ReadAllText(filePath);
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("CSV文件读取失败！error code=" + ex);
+                }
+            }
+            return "";
         }
     }
 }
